made everything linux friendly

+++ emitter.h
@@ -116,7 +116,7 @@ _create_event_handle (emitter_t *emitter, char *name, emitter_event_cb *cb, int 
   emitter_event_cb_t *cbh = malloc(sizeof(emitter_event_cb_t));
   cbh->cb = cb;
   cbh->once = once;
-  cbh->id = (int) cb;
+  cbh->id = (unsigned long) cb;
 
   event->cbs[event->cb_count++] = cbh;
