+++ test.c
@@ -81,10 +81,10 @@ main (void) {
   // assert that the listeners ids are the
   // address value of the function pointers
   //
-  assert((int) &on_foo1 == foo_listeners[0]->id);
-  assert((int) &on_foo2 == foo_listeners[1]->id);
-  assert((int) &on_bar1 == bar_listeners[0]->id);
-  assert((int) &on_bar2 == bar_listeners[1]->id);
+  assert((unsigned long)&on_foo1 == foo_listeners[0]->id);
+  assert((unsigned long) &on_foo2 == foo_listeners[1]->id);
+  assert((unsigned long) &on_bar1 == bar_listeners[0]->id);
+  assert((unsigned long) &on_bar2 == bar_listeners[1]->id);
 
   async(env, uv_default_loop()) {
     queue(env, emit_foo);
@@ -96,7 +96,7 @@ main (void) {
 
 static void
 on_foo1 (void *data) {
-  assert(42 == (int) data);
+  assert(42 == (unsigned long) data);
   printf("#1\n");
   usleep(5000); // this will block next event to be called
                 // for 5ms
@@ -106,7 +106,7 @@ static void
 on_foo2 (void *data) {
   if (NULL != data) {
     assert(0 == emitter_off(emitter, "foo", on_foo1));
-    assert(42 == (int) data);
+    assert(42 == (unsigned long) data);
     printf("#2\n");
     assert(0 == emitter_emit(emitter, "foo", NULL));
   } else {
